+++ SurvivorsPlus/Mercenary/MercenaryChanges.cs
@@ -1,5 +1,6 @@
 using R2API;
 using RoR2;
+using RoR2.EntityLogic;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
 
@@ -8,9 +9,11 @@ namespace SurvivorsPlus.Mercenary
     public class MercenaryChanges
     {
         private GameObject merc = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Merc/MercBody.prefab").WaitForCompletion();
-
+        private GameObject evisOverlapProjectile = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Merc/EvisOverlapProjectile.prefab").WaitForCompletion();
         public MercenaryChanges()
         {
+            GameObject.Destroy(evisOverlapProjectile.GetComponent<DelayedEvent>());
+
             ContentAddition.AddEntityState<EvisDashNux>(out _);
             ContentAddition.AddEntityState<EvisNux>(out _);
 
@@ -24,7 +27,7 @@ namespace SurvivorsPlus.Mercenary
             SkillLocator skillLocator2 = merc.GetComponent<SkillLocator>();
             skillLocator2.utility.skillFamily.variants[0].skillDef.skillDescriptionToken = "<style=cIsDamage>Stunning</style>. Dash forward, dealing <style=cIsDamage>400% damage</style>. If you hit an enemy, <style=cIsDamage>you can dash again</style>, up to <style=cIsDamage>3</style> total.";
             skillLocator2.utility.skillFamily.variants[1].skillDef.skillDescriptionToken = "<style=cIsDamage>Stunning</style>. Dash forward, dealing <style=cIsDamage>800% damage</style> and <style=cIsUtility>Exposing</style> enemies after<style=cIsUtility> 1 second</style>.";
-            skillLocator2.special.skillFamily.variants[0].skillDef.skillDescriptionToken = "Target all nearby enemies, attacking them for <style=cIsDamage>110% damage</style> repeatedly. <style=cIsUtility>You cannot be hit for the duration</style>.";
+            skillLocator2.special.skillFamily.variants[0].skillDef.skillDescriptionToken = "Target a nearby enemy, attacking them for <style=cIsDamage>150% damage</style> repeatedly. <style=cIsUtility>Exposing</style> on the final hit. <style=cIsUtility>You cannot be hit for the duration</style>.";
             skillLocator2.special.skillFamily.variants[0].skillDef.activationState = new EntityStates.SerializableEntityStateType(typeof(EvisDashNux));
         }
     }